+++ lib/sequent/support/database.rb
@@ -13,24 +13,8 @@ module Sequent
     class Database
       include ActiveRecord::Tasks
 
-      attr_reader :db_config
-
       def self.connect!(env)
-        db_config = read_config(env)
-        establish_connection(db_config)
-      end
-
-      def self.read_database_config(env)
-        fail ArgumentError, 'env is mandatory' unless env
-
-        DatabaseTasks.db_dir = Sequent.configuration.database_schema_directory unless defined?(Rails)
-        database_yml = File.join(Sequent.configuration.database_config_directory, 'database.yml')
-        config = YAML.safe_load(ERB.new(File.read(database_yml)).result, aliases: true)[env]
-        ActiveRecord::Base.configurations.resolve(config)
-      end
-
-      def self.read_config(env)
-        read_database_config(env).configuration_hash.with_indifferent_access
+        ActiveRecord::Base.establish_connection(env&.to_sym)
       end
 
       def self.create!(db_config)
@@ -41,15 +25,8 @@ module Sequent
         DatabaseTasks.drop(db_config)
       end
 
-      def self.establish_connection(db_config)
-        if Sequent.configuration.can_use_multiple_databases?
-          ActiveRecord::Base.configurations = db_config.stringify_keys
-          ActiveRecord::Base.connects_to database: {
-            Sequent.configuration.primary_database_role => Sequent.configuration.primary_database_key,
-          }
-        else
-          ActiveRecord::Base.establish_connection(db_config)
-        end
+      def self.establish_connection(db_config_or_env = Sequent.env&.to_sym)
+        ActiveRecord::Base.establish_connection(db_config_or_env)
       end
 
       def self.disconnect!
@@ -68,7 +45,7 @@ module Sequent
       end
 
       def self.drop_schema!(schema_name)
-        execute_sql "DROP SCHEMA if exists #{schema_name} cascade"
+        execute_sql "DROP SCHEMA IF EXISTS #{schema_name} CASCADE"
       end
 
       def self.with_search_path(search_path)