+++ spec/lib/sequent/generator/project_spec.rb
@@ -74,14 +74,12 @@ describe Sequent::Generator::Project do
     )
 
     Bundler.with_unbundled_env do
-      # Change default database configuration
-      Sequent.configuration.database_config_directory = File.join(path, 'db')
-      Database.write_database_yml_for_test(env: 'test', database_name: 'blog_with_special_symbols_test')
-
       system 'bash', '-cex', <<~SCRIPT
         cd blog-with_special-symbols
         export SEQUENT_ENV=test
         export BUNDLE_GEMFILE=./Gemfile
+        export PGUSER=sequent
+        export PGPASSWORD=sequent
 
         ruby_version=$(ruby -v | awk '{print $2}' | grep -o '^[0-9.]*')
         echo "$ruby_version" > .ruby-version